python-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
groq==0.4.2
motor==3.3.2
pymongo==4.6.1
slowapi==0.1.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pydantic==2.6.1
python-multipart==0.0.9
pinecone-client==3.0.2
pydantic-settings==2.1.0
openai==1.12.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
pytest==8.0.0
black==24.1.1
isort==5.13.2
mypy==1.8.0
